Update minesweeper.py
Move code for revealing the contents of a space on the board out of the function that runs when a board button is pressed, click(), into reveal_contents()

Add flag to tell when this particular space has been clicked or when the program wants to reveal the contents of this space. Handy because it can prevent certain functions, i.e., pool_fill() from being called more than once.